requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
gunicorn
psycopg2-binary
Werkzeug
flask-migrate
python-dotenv
psycopg2


